+++ Library/src/main/java/com/shopify/testify/ScreenshotRule.kt
@@ -70,6 +70,9 @@ import com.shopify.testify.internal.modification.HidePasswordViewModification
 import com.shopify.testify.internal.modification.HideScrollbarsViewModification
 import com.shopify.testify.internal.modification.HideTextSuggestionsViewModification
 import com.shopify.testify.internal.modification.SoftwareRenderViewModification
+import com.shopify.testify.internal.output.OutputFileUtility
+import com.shopify.testify.report.ReportSession
+import com.shopify.testify.report.Reporter
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
 import org.junit.rules.TestRule
@@ -91,13 +94,15 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
     private val activityClass: Class<T>,
     @IdRes private var rootViewId: Int = android.R.id.content,
     initialTouchMode: Boolean = false,
-    private val launchActivity: Boolean = true
+    private val launchActivity: Boolean = true,
+    enableReporter: Boolean = false
 ) : ActivityTestRule<T>(activityClass, initialTouchMode, launchActivity), TestRule {
 
     @LayoutRes
     private var targetLayoutId: Int = NO_ID
+
     @Suppress("MemberVisibilityCanBePrivate")
-    lateinit var testMethodName: String
+    open lateinit var testMethodName: String
     private lateinit var testClass: String
     private lateinit var testSimpleClassName: String
     private val hideCursorViewModification = HideCursorViewModification()
@@ -105,7 +110,7 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
     private val hideScrollbarsViewModification = HideScrollbarsViewModification()
     private val hideTextSuggestionsViewModification = HideTextSuggestionsViewModification()
     private val softwareRenderViewModification = SoftwareRenderViewModification()
-    private val testContext = getInstrumentation().context
+    internal val testContext = getInstrumentation().context
     private var assertSameInvoked = false
     private var espressoActions: EspressoActions? = null
     private var exactness: Float? = null
@@ -117,9 +122,17 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
     private var throwable: Throwable? = null
     private var viewModification: ViewModification? = null
     private var extrasProvider: ExtrasProvider? = null
+    @VisibleForTesting internal var reporter: Reporter? = null
+        private set
     private var orientationHelper = OrientationHelper(activityClass)
     private var exclusionRectProvider: ExclusionRectProvider? = null
     private val exclusionRects = HashSet<Rect>()
+
+    init {
+        if (enableReporter || TestifyFeatures.Reporter.isEnabled(getInstrumentation().context)) {
+            reporter = Reporter(getInstrumentation().targetContext, ReportSession(), OutputFileUtility())
+        }
+    }
 
     @Suppress("MemberVisibilityCanBePrivate")
     val testName: String
@@ -129,7 +142,7 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
         return Looper.getMainLooper().thread == Thread.currentThread()
     }
 
-    private val testNameComponents: TestName
+    internal val testNameComponents: TestName
         get() = TestName(testSimpleClassName, testMethodName)
 
     private val fullyQualifiedTestPath: String
@@ -276,12 +289,15 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
     }
 
     override fun apply(base: Statement, description: Description): Statement {
-        checkForScreenshotInstrumentationAnnotation(description)
-        applyExactness(description)
         espressoActions = null
         testSimpleClassName = description.testClass.simpleName
         testMethodName = description.methodName
         testClass = "${description.testClass?.canonicalName}#${description.methodName}"
+
+        reporter?.startTest(this, description)
+
+        checkForScreenshotInstrumentationAnnotation(description)
+        applyExactness(description)
         val testifyLayout: TestifyLayout? = description.getAnnotation(TestifyLayout::class.java)
         targetLayoutId = testifyLayout?.resolvedLayoutId ?: View.NO_ID
         return super.apply(ScreenshotStatement(base), description)
@@ -363,6 +379,8 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
 
         try {
             try {
+				reporter?.captureOutput(this)
+
                 if (isRunningOnUiThread()) {
                     throw NoScreenshotsOnUiThreadException()
                 }
@@ -496,13 +514,24 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
     }
 
     private inner class ScreenshotStatement constructor(private val base: Statement) : Statement() {
-
         override fun evaluate() {
-            assertSameInvoked = false
-            base.evaluate()
-            // Safeguard against accidentally omitting the call to `assertSame`
-            if (!assertSameInvoked) {
-                throw MissingAssertSameException()
+            try {
+                getInstrumentation()?.run {
+                    reporter?.identifySession(this)
+                }
+
+                assertSameInvoked = false
+                base.evaluate()
+                // Safeguard against accidentally omitting the call to `assertSame`
+                if (!assertSameInvoked) {
+                    throw MissingAssertSameException()
+                }
+                reporter?.pass()
+            } catch (throwable: Throwable) {
+                reporter?.fail(throwable)
+                throw throwable
+            } finally {
+                reporter?.endTest()
             }
         }
     }